Ticket Booking,Listing APIs

+++ backend/events/website/views.py
@@ -1,9 +1,9 @@
-from rest_framework import generics,status,permissions,filters
+from rest_framework import generics,status,permissions,filters,views
 from rest_framework.response import Response
 from rest_framework.pagination import LimitOffsetPagination
 from django_filters.rest_framework import DjangoFilterBackend
-from ..models import EventDetailModel,EventCategoryModel
-from .serializers import WebsiteEventDetailModelCreateSerializer,WebsiteEventDetailModelUpdateSerializer,WebsiteEventDetailModelListSerializer,WebsiteEventCategoryModelListSerializer
+from ..models import EventDetailModel,EventCategoryModel,EventAttendeeModel
+from .serializers import WebsiteEventDetailModelCreateSerializer,WebsiteEventDetailModelUpdateSerializer,WebsiteEventDetailModelListSerializer,WebsiteEventCategoryModelListSerializer,WebsiteEventAttendeeModelCreateSerializer,WebsiteEventAttendeeModelListSerializer
 from ..serializers_utils import UtilsEventDetailModelSerializer
 
 
@@ -104,14 +104,60 @@ class WebsiteRecipeModelGenericAPIView(generics.GenericAPIView):
     def delete(self,request,id):
         try:
             queryset= self.queryset.get(pk=id)
-            if request.user.pk == queryset.organiser.pk:
+            if request.user.pk != queryset.organiser.pk and queryset.capacity == queryset.available:
+                return Response({"message" : "You are not authorized to update this event"}, status=status.HTTP_400_BAD_REQUEST)
+            elif queryset.capacity != queryset.available:
+                return Response({"message" : "Some tickets are already booked, Contact Admin to Update !!!"}, status=status.HTTP_400_BAD_REQUEST)
+            else:
                 queryset.delete()
                 return Response({"message" : "Deleted Successfully"}, status=status.HTTP_200_OK)
-            else:
-                return Response({"message" : "Not Allowed"}, status=status.HTTP_400_BAD_REQUEST)
         except Exception as e:
             return Response({"message" : f"Something went wrong-->{e}"}, status=status.HTTP_400_BAD_REQUEST)
 
 class WebsiteEventCategoryModelListAPIView(generics.ListAPIView):
     queryset = EventCategoryModel.objects.all()
     serializer_class = WebsiteEventCategoryModelListSerializer
+
+class WebsiteEventDetailModelPrivacyUpdateAPIView(views.APIView):
+    permission_classes=[permissions.IsAuthenticated]
+
+    def put(self,request,id):
+        try:
+            event_instance = EventDetailModel.objects.get(id=id)
+            if self.request.user.pk != event_instance.organiser.pk:
+                return Response({"message" : "You are not authorized to update this event"}, status=status.HTTP_400_BAD_REQUEST)
+            elif event_instance.available != event_instance.capacity:
+                return Response({"message" : "Some tickets are already booked, Contact Admin to Update !!!"}, status=status.HTTP_400_BAD_REQUEST)
+            else:
+                event_instance.is_private = not event_instance.is_private
+                event_instance.save()
+                return Response({"message" : "Privacy Updated Successfully"}, status=status.HTTP_200_OK)
+        except Exception as e:
+            return Response({"message" : f"Something went wrong: {e}"}, status=status.HTTP_400_BAD_REQUEST)
+
+
+class WebsiteEventAttendeeModelCreateAPIView(generics.CreateAPIView):
+    queryset = EventAttendeeModel.objects.all()
+    serializer_class=WebsiteEventAttendeeModelCreateSerializer
+    permission_classes=[permissions.IsAuthenticated]
+
+    def create(self, request):
+        try:
+            serializer  = self.serializer_class(data=request.data,context={'request':request})
+            if serializer.is_valid():
+                serializer.save()
+                return Response({"message" : "Event Created Successfully"}, status=status.HTTP_200_OK)
+            else:
+                return Response({"message" : serializer.errors}, status=status.HTTP_400_BAD_REQUEST) 
+        except Exception as e:
+            return Response({"message" : f"something went wrong, {e}"}, status=status.HTTP_400_BAD_REQUEST)
+
+
+class WebsiteEventAttendeeModelListAPIView(generics.ListAPIView):
+    serializer_class = WebsiteEventAttendeeModelListSerializer
+    permission_classes=[permissions.IsAuthenticated]
+
+
+    def get_queryset(self):
+        return self.request.user.EventAttendeeModel_user.all().order_by('-created_at')
+
